universal9830: defconfig: Disable Samsung Software WatchDog (HACK)
Some components are not happy when SDP and DDAR are disabled and
they cause kernel to panic due to "Software Watchdog Timeout".

Looking into /vendor/etc/init/hw/init.samsung.bsp.rc, it is found
that Samsung intends that watchdog to be disabled in custom ROMs:

	on property:ro.boot.verifiedbootstate=orange
	    write /dev/watchdog "V"

Thus, this change disables it altogether in kernel to make things
smoother.